refactor: reorganie api directory structure
Rename /api/routes to /api/endpoints
